V1 Frontend done
Need to figure out this error to use GenerateProjectPlan and inject its output into Generate Task which is connected to Jira REST api 3. ***ERROR***" Module not found: Error: Can't resolve 'js-tiktoken/lite' in 'C:\Users\Alex School\Desktop\git\Atlassian-Forge\smart-plan\node_modules\langchain\dist\util'
Did you mean 'lite.js'?
BREAKING CHANGE: The request 'js-tiktoken/lite' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request."